+++ src/domain/GameScreen/Activity/Kings/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import CardImage from '../CardImage'
+import CardImage from '../../CardImage'
 
 import kingsStyle from './style.module.scss'
 
@@ -7,8 +7,8 @@ const Kings = () => {
 
   const kings = ['srce_kralj', 'pik_kralj', 'kara_kralj', 'kriz_kralj']
 
-  const Kings = kings.map(king =>
-    <CardImage card={king} />
+  const Kings = kings.map((king, index) =>
+    <CardImage card={king} key={index} />
   )
 
   return (
