+++ src/domain/GameScreen/Activity/Kings/style.module.scss
