+++ src/domain/GameScreen/Activity/Talon/index.tsx
@@ -1,14 +1,14 @@
 import React from 'react'
 
-import CardImage from '../CardImage'
+import CardImage from '../../CardImage'
 
 import talonStyle from './style.module.scss'
 
 const Talon = ({cards}) => {
   
-  const Cards = cards.map(part =>
-    <div className={talonStyle.part}>
-      { part.map(card => <CardImage card={card} />) }
+  const Cards = cards.map((part, index) =>
+    <div className={talonStyle.part} key={index}>
+      { part.map((card, index) => <CardImage card={card} key={index} />) }
     </div>
   )
