+++ src/domain/GameScreen/Activity/Talon/style.module.scss
